+++ instrumentation/net/http/otelhttp/config.go
@@ -29,6 +29,7 @@ type config struct {
 	PublicEndpointFn  func(*http.Request) bool
 	ReadEvent         bool
 	WriteEvent        bool
+	WithoutRouteAttr  bool
 	Filters           []Filter
 	SpanNameFormatter func(string, *http.Request) string
 	ClientTrace       func(context.Context) *httptrace.ClientTrace
@@ -161,6 +162,13 @@ func WithMessageEvents(events ...Event) Option {
 				c.WriteEvent = true
 			}
 		}
+	})
+}
+
+// WithoutMetricRouteAttribute returns an [Option] that disables recording the HTTP route path as a metric attribute.
+func WithoutMetricRouteAttribute() Option {
+	return optionFunc(func(c *config) {
+		c.WithoutRouteAttr = true
 	})
 }