+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 
 - Add environment variables propagation carrier in `go.opentelemetry.io/contrib/propagators/envcar`. (#8442)
+- Add `WithMetricRouteAttribute` option in `go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp` (#8632)
 
 ### Changed